backend/requirements.txt
flask
flask-cors
boto3
python-dotenv
assemblyai
